requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==4.2.16
django-environ==0.11.2
environ==1.0
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
pillow==11.0.0
pydantic==2.10.1
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.2
trycourier==6.2.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
user-agent==0.1.10
